+++ src/main/java/pl/shopmatelist/shopmatelist/services/ProductsService.java
@@ -1,15 +1,16 @@
 package pl.shopmatelist.shopmatelist.services;
 
 import lombok.AllArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
 import pl.shopmatelist.shopmatelist.dto.ProductsDTO;
 import pl.shopmatelist.shopmatelist.entity.Products;
+import pl.shopmatelist.shopmatelist.exceptions.IllegalArgumentException;
+import pl.shopmatelist.shopmatelist.exceptions.ProductNotFoundException;
 import pl.shopmatelist.shopmatelist.mapper.ProductsMapper;
 import pl.shopmatelist.shopmatelist.repository.ProductsRepository;
 
 import java.util.List;
-import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Service
@@ -20,38 +21,53 @@ public class ProductsService {
     private final ProductsMapper productsMapper;
 
 
-    public ProductsDTO findById(Long id){
+    public ProductsDTO findById(Long id) {
         Optional<Products> optionalProduct = productsRepository.findById(id);
-        if(optionalProduct.isPresent()){
+        if (optionalProduct.isPresent()) {
             Products product = optionalProduct.get();
             return productsMapper.toDTO(product);
         }
-        throw new NoSuchElementException();
+        throw new ProductNotFoundException("Nie ma produktu o podanym id:  " + id);
     }
 
-    public List<ProductsDTO> findAll(){
+    public List<ProductsDTO> findAll() {
         List<Products> products = productsRepository.findAll();
         return productsMapper.toDtoList(products);
     }
 
 
     public ProductsDTO createProducts(ProductsDTO productsDTO) {
-        Products products = productsMapper.toEntity(productsDTO);
-        Products savedProducts = productsRepository.save(products);
-        return productsMapper.toDTO(savedProducts);
+        List<Products> products = productsRepository.findAll();
+        Products product = productsMapper.toEntity(productsDTO);
+        boolean exist = products.stream().anyMatch(productInDB -> productInDB.getProductName().equals(product.getProductName().toLowerCase()) || productInDB.getProductId().equals(product.getProductId()));
+        if (exist) {
+            throw new IllegalArgumentException("Produkt o nazwie: " + product.getProductName() + " już istnieje");
+        } else {
+            Products savedProducts = productsRepository.save(product);
+            return productsMapper.toDTO(savedProducts);
+        }
+
     }
 
-    public void deleteById(Long id){
+    public void deleteById(Long id) {
+
+        Optional<Products> productToDelete = productsRepository.findById(id);
+        if (productToDelete.isEmpty()) {
+            throw new ProductNotFoundException("Nie ma produktu o id: " + id + " w bazie");
+        }
         productsRepository.deleteById(id);
+
     }
 
-    public ProductsDTO update(ProductsDTO productsDTO){
+    public ProductsDTO update(ProductsDTO productsDTO) {
+        if (productsDTO.getProductId() == null) {
+            throw new IllegalArgumentException("ID produktu nie może byc puste!");
+        }
+
         Products product = productsMapper.toEntity(productsDTO);
         Products updatedProduct = productsRepository.save(product);
         return productsMapper.toDTO(updatedProduct);
     }
-
-
 
 
 }
